+++ app/Http/Controllers/TopBannerController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Http\Requests\TopBannerRequest;
 use Yajra\DataTables\Facades\DataTables;
 use App\Models\{Tag, TopBanner};
+use Illuminate\Support\Facades\Auth;
 
 class TopBannerController extends Controller
 {
@@ -15,8 +16,12 @@ class TopBannerController extends Controller
     // Display a listing of the resource.
     public function index()
     {
-        $total_top_banner = TopBanner::count();
-        return view('admin.top_banners.index', compact(['total_top_banner']));
+        if(Auth::guard('admin')->user()->can('top-banners.index')){
+            $total_top_banner = TopBanner::count();
+            return view('admin.top_banners.index', compact(['total_top_banner']));
+        }else{
+            return redirect()->route('admin.dashboard')->with('error','You have no rights for this action!');
+        }
     }
 
     // Load all top banners helping with AJAX Datatable
@@ -46,7 +51,11 @@ class TopBannerController extends Controller
             })
             ->addColumn('status', function ($row){
                 $isChecked = ($row->status == 1) ? 'checked' : '';
-                return '<div class="form-check form-switch"><input class="form-check-input" type="checkbox" role="switch" onchange="changeStatus(\''.encrypt($row->id).'\')" id="statusBtn" '.$isChecked.'></div>';
+                if(Auth::guard('admin')->user()->can('top-banners.status')){
+                    return '<div class="form-check form-switch"><input class="form-check-input" type="checkbox" role="switch" onchange="changeStatus(\''.encrypt($row->id).'\')" id="statusBtn" '.$isChecked.'></div>';
+                }else{
+                    return '<div class="form-check form-switch"><input class="form-check-input" type="checkbox" role="switch" id="statusBtn" '.$isChecked.' disabled></div>';
+                }
             })
             ->addColumn('tag', function ($row){
                 $tag_name = (isset($row['tag']['name'])) ? $row['tag']['name'] : '';
@@ -54,8 +63,17 @@ class TopBannerController extends Controller
             })
             ->addColumn('actions',function($row){
                 $action_html = '';
-                $action_html .= '<a href="'.route('top-banners.edit',encrypt($row->id)).'" class="btn btn-sm custom-btn me-1"><i class="bi bi-pencil"></i></a>';
-                $action_html .= '<a onclick="deleteTopBanner(\''.encrypt($row->id).'\')" class="btn btn-sm btn-danger me-1"><i class="bi bi-trash"></i></a>';
+                if(Auth::guard('admin')->user()->can('top-banners.edit')){
+                    $action_html .= '<a href="'.route('top-banners.edit',encrypt($row->id)).'" class="btn btn-sm custom-btn me-1"><i class="bi bi-pencil"></i></a>';
+                }else{
+                    $action_html .= '- ';
+                }
+
+                if(Auth::guard('admin')->user()->can('top-banners.destroy')){
+                    $action_html .= '<a onclick="deleteTopBanner(\''.encrypt($row->id).'\')" class="btn btn-sm btn-danger me-1"><i class="bi bi-trash"></i></a>';
+                }else{
+                    $action_html .= '- ';
+                }
                 return $action_html;
             })
             ->rawColumns(['status','actions','image','tag','day'])
@@ -66,8 +84,12 @@ class TopBannerController extends Controller
     // Show the form for creating a new resource.
     public function create()
     {
-        $tags = Tag::get();
-        return view('admin.top_banners.create', compact(['tags']));
+        if(Auth::guard('admin')->user()->can('top-banners.create')){
+            $tags = Tag::get();
+            return view('admin.top_banners.create', compact(['tags']));
+        }else{
+            return redirect()->route('admin.dashboard')->with('error','You have no rights for this action!');
+        }
     }
 
     // Store a newly created resource in storage.
@@ -113,9 +135,13 @@ class TopBannerController extends Controller
     public function edit($id)
     {
         try{
-            $top_banner = TopBanner::find(decrypt($id));
-            $tags = Tag::all();
-            return view('admin.top_banners.edit', compact(['top_banner','tags']));
+            if(Auth::guard('admin')->user()->can('top-banners.edit')){
+                $top_banner = TopBanner::find(decrypt($id));
+                $tags = Tag::all();
+                return view('admin.top_banners.edit', compact(['top_banner','tags']));
+            }else{
+                return redirect()->route('admin.dashboard')->with('error','You have no rights for this action!');
+            }
         }catch (\Throwable $th){
             return redirect()->back()->with('error', 'Oops, Something went wrong!');
         }
